refactor(tasks): extract exec-tools + maintenance, thin the index
Completes the tasks/index split (WS6). The ~330-line block of 11 MCP exec tools
and the background-maintenance timers move out of activate() into lib modules:

- lib/exec-tools.ts — registerTaskExecTools(ctx): all 11 tools (list/get/create/
  move/block/complete/log_progress/set_dependency/update/delete/assign) plus the
  buildChecklistNudge advisory. They share the route helpers (edit-guard, search
  indexing, schemas) and import them from the same lib modules.
- lib/maintenance.ts — startMaintenance(ctx)/stopMaintenance(): the hourly
  auto-archive + periodic old-task-cleanup timers. startMaintenance clears any
  prior timers first, preserving the clear-before-start re-entrancy contract for
  hot reload; onShutdown calls stopMaintenance().

activate() is now the memory-plugin thin-index shape: register the search content
type, registerTaskExecTools(ctx), startMaintenance(ctx), and the 4 health checks.

index.ts: 539 -> 138 lines (1,089 -> 138 across the two tasks PRs). Pure
relocation — the optional behavioral redesigns the audit lists (REST/MCP guard
inconsistency, identifier-fallback dedup, dead .catch noise, postJson helper)
are deliberately NOT taken; noted for follow-up.

Verified: typecheck + lint clean; tasks suite 206/0; full suite 5123/0; binary
build; isolated-home boot smoke (Tasks activates; POST /api/exec-tools/
bakin_exec_tasks_list -> 200 confirms registerTaskExecTools wired; GET
/api/plugins/tasks/ -> 200).
